Routing
I took the liberty and started forming an outline for the routes, i also tried to make some css formating more friendly.
