+++ .gitignore
@@ -8,3 +8,4 @@ server
 package-lock.json
 server.exe
 package.json
+package.json
