+++ public/global.css
@@ -1,9 +1,6 @@
 html, body {
-	position: relative;
 	width: 100%;
 	height: 100%;
-	padding: 0px;
-	margin:	0px
 }
 
 body {
