+++ src/components/Home.svelte
@@ -1,0 +1,6 @@
+<script>
+</script>
+  
+  <div>
+    <h3>This is HomePage</h3>
+  </div>
